fix: get_mac_address only using getifaddrs Also, updated fmt to latest fixes #85

+++ src/core/src/platforms/all/docker/docker/formatters.hpp
@@ -58,7 +58,7 @@ public:
         container.id,
         container.name,
         container.image,
-        container.status,
+        (int)container.status,
         container.ports,
         container.mounts,
         container.devices,
